# openapi.yaml
openapi: 3.0.0
info:
  title: MarketMind AI - Financial Research Platform API
  description: >
    Production-ready REST API contract for the MarketMind AI Financial Research Platform.
    Provides endpoints for user authentication, stock listing, company profiling,
    analyst consensus tracking, stock pricing charts, earnings call transcripts,
    automated research runs, semantic vector searches (RAG), watchlists, scheduled cron jobs, and alerts.
  version: 1.0.0
servers:
  - url: https://api.marketmind.ai/v1
    description: Production API Server
paths:
...
  /stocks:
    get:
      summary: List and search tracked stock symbols
      tags:
        - Stocks
      security:
        - BearerAuth: []
      parameters:
        - name: query
          in: query
          description: Search term matching ticker or company name (supports fuzzy match)
          required: false
          schema:
            type: string
            example: App
        - name: sector
          in: query
          required: false
          schema:
            type: string
            example: Technology
        - name: exchange
          in: query
          required: false
          schema:
            type: string
            example: NASDAQ
      responses:
        '200':
          description: A list of stocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stock'
  /stocks/{id}:
    get:
      summary: Get details of a specific stock
      tags:
        - Stocks
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Stock found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockDetail'
        '404':
          description: Stock not found
  /stocks/{id}/profile:
    get:
      summary: Retrieve detailed Yahoo Finance profile metadata
      tags:
        - Stocks
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Company profile fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyProfile'
  /stocks/{id}/consensus:
    get:
      summary: Fetch external analyst ratings consensus and target pricing
      tags:
        - Stocks
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Consensus details fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalystConsensus'
  /stocks/{symbol}/prices:
    get:
      summary: Fetch historical prices (time-series chart data)
      tags:
        - Stocks
      security:
        - BearerAuth: []
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
            example: AAPL
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date
            example: '2026-01-01'
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date
            example: '2026-06-05'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            example: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            example: 0
      responses:
        '200':
          description: Price logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedStockPriceResponse'
  /stocks/{symbol}/fundamentals:
    get:
      summary: Fetch corporate fundamentals history
      tags:
        - Stocks
      security:
        - BearerAuth: []
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
            example: AAPL
        - name: period_type
          in: query
          required: false
          schema:
            type: string
            enum: [annual, quarterly, trailing_twelve_months]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            example: 10
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            example: 0
      responses:
        '200':
          description: Fundamentals list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCompanyFundamentalResponse'
...
  /watchlists:
    get:
      summary: List user watchlists
      tags:
        - Watchlists
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Watchlists list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Watchlist'
    post:
      summary: Create a new watchlist folder
      tags:
        - Watchlists
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Tech Watch
                description:
                  type: string
                  example: High-cap growth stocks
      responses:
        '201':
          description: Watchlist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Watchlist'
  /watchlists/{id}/items:
    post:
      summary: Add a stock ticker to a watchlist
      tags:
        - Watchlists
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - stock_id
              properties:
                stock_id:
                  type: string
                  format: uuid
                sort_order:
                  type: integer
                  default: 0
      responses:
        '201':
          description: Stock linked to watchlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistItem'
  /watchlists/{id}/items/{stock_id}:
    delete:
      summary: Remove a stock from a watchlist
      tags:
        - Watchlists
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: stock_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Stock removed from list
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Stock:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ticker:
          type: string
        name:
          type: string
        exchange:
          type: string
        sector:
          type: string
        industry:
          type: string
        is_active:
          type: boolean
    StockDetail:
      allOf:
        - $ref: '#/components/schemas/Stock'
        - type: object
          properties:
            profile:
              $ref: '#/components/schemas/CompanyProfile'
            consensus:
              $ref: '#/components/schemas/AnalystConsensus'
    CompanyProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        headquarters:
          type: string
        ceo:
          type: string
        employees:
          type: integer
        website:
          type: string
        founded_year:
          type: integer
        market_cap:
          type: number
        shares_outstanding:
          type: number
        updated_at:
          type: string
          format: date-time
    AnalystConsensus:
      type: object
      properties:
        id:
          type: string
          format: uuid
        buy_count:
          type: integer
        hold_count:
          type: integer
        sell_count:
          type: integer
        average_target_price:
          type: number
        updated_at:
          type: string
          format: date-time
    StockPrice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        price_date:
          type: string
          format: date
        open_price:
          type: number
        high_price:
          type: number
        low_price:
          type: number
        close_price:
          type: number
        volume:
          type: integer
        adjusted_close:
          type: number
    CompanyFundamental:
      type: object
      properties:
        id:
          type: string
          format: uuid
        report_date:
          type: string
          format: date
        period_type:
          type: string
        revenue:
          type: number
        net_income:
          type: number
        eps:
          type: number
        ebitda:
          type: number
        assets:
          type: number
        liabilities:
          type: number
        cash_flow:
          type: number
        metadata:
          type: object
...
    Watchlist:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
    WatchlistItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        watchlist_id:
          type: string
          format: uuid
        stock_id:
          type: string
          format: uuid
        added_at:
          type: string
          format: date-time
        sort_order:
          type: integer
...
    PaginatedStockPriceResponse:
      type: object
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StockPrice'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    PaginatedCompanyFundamentalResponse:
      type: object
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CompanyFundamental'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer